Create a standalone drinking session window component
